+++ complejo-deportivo/src/main/java/Mixins/Caracteristica.scala
@@ -6,7 +6,7 @@ import org.joda.time.DateTime
 trait Caracteristica extends Cancha{
 
   override 
-  def calcularExtra(dia : DateTime, inicio : Int, fin : Int, pb : Double) : Double = {
+  def calcularExtra(dia : DateTime, inicio : Double, fin : Double, pb : Double) : Double = {
 		  (this.extraPorCaracteristica(pb) + super.calcularExtra(dia, inicio, fin, pb))
   }
